requirements.txt
# Simple Newsletter Generator Requirements


# Web scraping with browser automation
playwright==1.40.0

# AI integration for article summarization
mistralai==0.4.2

# Environment variables for API keys
python-dotenv==1.0.0

# HTTP requests (backup for simple requests)
requests==2.31.0

# HTML parsing (helpful for cleaning text)
beautifulsoup4==4.12.2

# Advanced NLP Libraries for Smart Content Analysis
nltk==3.8.1          # Natural Language Toolkit for text processing
spacy>=3.7.0         # Industrial-strength NLP library
textblob==0.17.1     # Simple API for diving into common NLP tasks
scikit-learn>=1.3.0  # Machine learning for text classification and analysis
numpy>=1.24.0        # Numerical computing for NLP operations
pandas>=2.0.0        # Data manipulation and analysis

# Zero-Shot Classification with Transformers
transformers>=4.36.0  # Hugging Face transformers for BERT/DistilBERT
torch>=2.0.0         # PyTorch backend for transformers


